Pull extractLines and readTextChunks to their own files
